+++ paparazzi-gradle-plugin/src/main/java/app/cash/paparazzi/gradle/PrepareResourcesTask.kt
@@ -15,83 +15,91 @@
  */
 package app.cash.paparazzi.gradle
 
-import com.android.build.gradle.BaseExtension
-import com.android.ide.common.symbols.getPackageNameFromManifest
 import org.gradle.api.DefaultTask
-import org.gradle.api.Project
+import org.gradle.api.file.ConfigurableFileCollection
+import org.gradle.api.file.Directory
 import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.file.RegularFileProperty
+import org.gradle.api.provider.Property
 import org.gradle.api.tasks.CacheableTask
+import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.InputDirectory
+import org.gradle.api.tasks.InputFiles
 import org.gradle.api.tasks.OutputFile
 import org.gradle.api.tasks.PathSensitive
 import org.gradle.api.tasks.PathSensitivity
 import org.gradle.api.tasks.TaskAction
-import java.io.File
 
 @CacheableTask
-open class PrepareResourcesTask : DefaultTask() {
-  @get:InputDirectory
-  @get:PathSensitive(PathSensitivity.RELATIVE)
-  internal val mergeResourcesOutput: DirectoryProperty = project.objects.directoryProperty()
+abstract class PrepareResourcesTask : DefaultTask() {
+  @get:Input
+  abstract val packageName: Property<String>
 
   @get:InputDirectory
   @get:PathSensitive(PathSensitivity.RELATIVE)
-  internal val mergeAssetsOutput: DirectoryProperty = project.objects.directoryProperty()
+  abstract val mergeResourcesOutput: DirectoryProperty
+
+  @get:Input
+  abstract val targetSdkVersion: Property<String>
+
+  @get:Input
+  abstract val compileSdkVersion: Property<String>
+
+  @get:InputDirectory
+  @get:PathSensitive(PathSensitivity.RELATIVE)
+  abstract val mergeAssetsOutput: DirectoryProperty
+
+  @get:InputDirectory
+  @get:PathSensitive(PathSensitivity.RELATIVE)
+  abstract val platformDataRoot: DirectoryProperty
+
+  @get:Input
+  abstract val nonTransitiveRClassEnabled: Property<Boolean>
+
+  @get:InputFiles
+  @get:PathSensitive(PathSensitivity.NONE)
+  abstract val artifactFiles: ConfigurableFileCollection
 
   @get:OutputFile
-  internal val paparazziResources: RegularFileProperty = project.objects.fileProperty()
+  abstract val paparazziResources: RegularFileProperty
+
+  private val projectDirectory = project.layout.projectDirectory
 
   @TaskAction
   fun writeResourcesFile() {
     val out = paparazziResources.get().asFile
     out.delete()
+
+    val mainPackage = packageName.get()
+    val resourcePackageNames = if (nonTransitiveRClassEnabled.get()) {
+      "$mainPackage," + artifactFiles.files.joinToString(",") { file ->
+        file.useLines { lines -> lines.first() }
+      }
+    } else {
+      mainPackage
+    }
+
     out.bufferedWriter()
-        .use {
-          it.write(project.packageName())
-          it.newLine()
-          it.write(project.relativePath(mergeResourcesOutput.get().asFile.path))
-          it.newLine()
-          it.write(project.targetSdkVersion())
-          it.newLine()
-          // Use compileSdkVersion for system framework resources.
-          it.write("platforms/android-${project.compileSdkVersion()}/")
-          it.newLine()
-          it.write(project.relativePath(mergeAssetsOutput.get().asFile.path))
-          it.newLine()
-        }
+      .use {
+        it.write(mainPackage)
+        it.newLine()
+        it.write(projectDirectory.relativize(mergeResourcesOutput.get()))
+        it.newLine()
+        it.write(targetSdkVersion.get())
+        it.newLine()
+        // Use compileSdkVersion for system framework resources.
+        it.write("platforms/android-${compileSdkVersion.get()}/")
+        it.newLine()
+        it.write(projectDirectory.relativize(mergeAssetsOutput.get()))
+        it.newLine()
+        it.write(platformDataRoot.get().asFile.invariantSeparatorsPath)
+        it.newLine()
+        it.write(resourcePackageNames)
+        it.newLine()
+      }
   }
 
-  private fun Project.packageName(): String {
-    val androidExtension = extensions.getByType(BaseExtension::class.java)
-    androidExtension.sourceSets
-        .map { it.manifest.srcFile }
-        .filter { it.exists() }
-        .forEach {
-          return getPackageNameFromManifest(it)
-        }
-    throw IllegalStateException("No source sets available")
-  }
-
-  private fun Project.compileSdkVersion(): String {
-    val androidExtension = extensions.getByType(BaseExtension::class.java)
-    return androidExtension.compileSdkVersion!!.substringAfter(
-        "android-", DEFAULT_COMPILE_SDK_VERSION.toString()
-    )
-  }
-
-  private fun Project.targetSdkVersion(): String {
-    val androidExtension = extensions.getByType(BaseExtension::class.java)
-    return androidExtension.defaultConfig.targetSdkVersion?.apiLevel?.toString()
-        ?: DEFAULT_COMPILE_SDK_VERSION.toString()
-  }
-
-  private fun Project.sdkFolder(): File {
-    val androidExtension = extensions.getByType(BaseExtension::class.java)
-    return androidExtension.sdkDirectory
-  }
-
-  companion object {
-    const val DEFAULT_COMPILE_SDK_VERSION = 29
+  private fun Directory.relativize(child: Directory): String {
+    return asFile.toPath().relativize(child.asFile.toPath()).toFile().invariantSeparatorsPath
   }
 }