+++ paparazzi/src/main/java/app/cash/paparazzi/HtmlReportWriter.kt
@@ -57,7 +57,7 @@ import javax.imageio.ImageIO
  * paparazzi.js
  * ```
  */
-internal class HtmlReportWriter(
+class HtmlReportWriter @JvmOverloads constructor(
   private val runName: String = defaultRunName(),
   private val rootDirectory: File = File("build/reports/paparazzi"),
   snapshotRootDirectory: File = File("src/test/snapshots")
@@ -96,13 +96,15 @@ internal class HtmlReportWriter(
       }
 
       override fun close() {
+        if (hashes.isEmpty()) return
+
         val shot = if (hashes.size == 1) {
           val original = File(imagesDirectory, "${hashes[0]}.png")
           if (isRecording) {
             val goldenFile = File(goldenImagesDirectory, snapshot.toFileName("_", "png"))
             original.copyTo(goldenFile, overwrite = true)
           }
-          snapshot.copy(file = rootDirectory.toPath().relativize(original.toPath()).toString())
+          snapshot.copy(file = original.toJsonPath())
         } else {
           val hash = writeVideo(hashes, fps)
 
@@ -123,7 +125,7 @@ internal class HtmlReportWriter(
               original.copyTo(goldenFile)
             }
           }
-          snapshot.copy(file = rootDirectory.toPath().relativize(original.toPath()).toString())
+          snapshot.copy(file = original.toJsonPath())
         }
 
         shots += shot
@@ -258,6 +260,7 @@ internal class HtmlReportWriter(
   private fun File.writeAtomically(bufferedImage: BufferedImage) {
     val tmpFile = File(parentFile, "$name.tmp")
     ImageIO.write(bufferedImage, "PNG", tmpFile)
+    delete()
     tmpFile.renameTo(this)
   }
 
@@ -268,8 +271,11 @@ internal class HtmlReportWriter(
         .use { sink ->
           sink.writerAction()
         }
+    delete()
     tmpFile.renameTo(this)
   }
+
+  private fun File.toJsonPath(): String = relativeTo(rootDirectory).invariantSeparatorsPath
 }
 
 internal fun defaultRunName(): String {